Finetuning unified again and new ProgressMonitor
- ProgressMonitor now internally is a stack of Tasks, each Task has its
own 0-1 progress range which is mapped to some range interval of the
parent task

- MaskEmbedder and MaskExtractor now allow to embed masks as overlay in
image stacks. That way Finetuning is unified again, because all training
samples are simply images with labels in the overlay. This allows easy
storage of all data relevant for training in a single TIFF file. Since
ImageRois are not easily handled the new plugins MaskEmbedder and
MaskExtractor do the embedding and extraction.